% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsize.R
\name{betas}
\alias{betas}
\title{Compute standardized coefficients}
\usage{
betas(model, ...)
}
\arguments{
\item{obj}{of class \code{\link[lme4]{merMod-class}}}
}
\value{
a data.frame with the original coefficients and the beta coefficients
}
\description{
Compute the standardized coefficients, betas, for a mixed model.
}
\examples{
 methods(class="pamlpowermix")
}
\author{
Marcello Gallucci
}
