% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{Design}
\alias{Design}
\title{Design R6 Class}
\description{
The R6 object that defines the sample, the design matrix and runs the simulations for power calculations
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Design-new}{\code{Design$new()}}
\item \href{#method-Design-print}{\code{Design$print()}}
\item \href{#method-Design-check}{\code{Design$check()}}
\item \href{#method-Design-size}{\code{Design$size()}}
\item \href{#method-Design-create_model}{\code{Design$create_model()}}
\item \href{#method-Design-update_model}{\code{Design$update_model()}}
\item \href{#method-Design-one_sample}{\code{Design$one_sample()}}
\item \href{#method-Design-simulate}{\code{Design$simulate()}}
\item \href{#method-Design-options}{\code{Design$options()}}
\item \href{#method-Design-clone}{\code{Design$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-new"></a>}}
\if{latex}{\out{\hypertarget{method-Design-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Design object. No arguments are required.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-print"></a>}}
\if{latex}{\out{\hypertarget{method-Design-print}{}}}
\subsection{Method \code{print()}}{
Print some information regarding the design being set up
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-check"></a>}}
\if{latex}{\out{\hypertarget{method-Design-check}{}}}
\subsection{Method \code{check()}}{
Checks some crucial parameters of the object being defined.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$check()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-size"></a>}}
\if{latex}{\out{\hypertarget{method-Design-size}{}}}
\subsection{Method \code{size()}}{
Define new sample size parameters. It accepts clusters name as arguments in the form
\code{obj$size(cluster1='integer', cluster2='integer', ...)}. Not necessary to use when \code{obj$simulate()} is used.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$size(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-create_model"></a>}}
\if{latex}{\out{\hypertarget{method-Design-create_model}{}}}
\subsection{Method \code{create_model()}}{
Create a root model as a basis of the power simulations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$create_model(family = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{family}}{a \code{\link[stats]{stats::family}}  that defines the type of model to be estimated. Any family accepted by
\code{\link[lme4]{lme4::glmer}} is accepted. If not defined (default) a linear model is assumed. If random coefficients
are defined (such as \code{(x|cluster)} in the formula), a mixed model is assumed, eastimated with \code{\link[lme4]{lme4::lmer}}. If no random effects are defined, a general(ized) linear
model is estimated with \code{\link[stats]{lme}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-update_model"></a>}}
\if{latex}{\out{\hypertarget{method-Design-update_model}{}}}
\subsection{Method \code{update_model()}}{
Update a previously defined model with new coeffcients or new sample sizes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$update_model(
  data = NULL,
  fixed = NULL,
  varcorr = NULL,
  sigma = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a new dataframe to use to estimate the model}

\item{\code{fixed}}{a new set of fixed coefficient defined as \code{fixed=list(varname1=number,varname2=number)}}

\item{\code{random}}{a new set of random coefficient defined as \code{random=list(acluster=list(varname1=number,varname2=number))}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-one_sample"></a>}}
\if{latex}{\out{\hypertarget{method-Design-one_sample}{}}}
\subsection{Method \code{one_sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$one_sample(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-Design-simulate}{}}}
\subsection{Method \code{simulate()}}{
Define a simulation experiment. It accepts several arguments, depending to the type of experiment the user wishes to use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$simulate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{any cluster variable defined with \code{obj$clusters} with a required size, passed as \verb{obj$simulate(aclustername='integer'} where
\code{integer} is the number of clusters required.}

\item{\code{Nsim}}{the number of simulations (repetitions) for each cell of the simulation experimental design.}

\item{\code{fixed}}{a named list of of expected parameters of the form \verb{(fixed=list('(Intercept)'='number',aname='number',...)}.
Parameters not defined here are set accordingly to the formula passed with \code{obj$formula<-astring} method.}

\item{\code{random}}{a named list of of expected parameters of the form \verb{(random=list(acluster=list('(Intercept)'='number',aname='number',...))}.
Parameters not defined here are set accordingly to the formula passed with \code{obj$formula<-astring} method.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-options"></a>}}
\if{latex}{\out{\hypertarget{method-Design-options}{}}}
\subsection{Method \code{options()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$options(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Design-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
