% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{get_sample}
\alias{get_sample}
\title{Create a dataset based on a data generating model}
\usage{
get_sample(info)
}
\arguments{
\item{vars}{a list defining variables characteristics}

\item{clusters}{a list defining clusters carateristics}

\item{structure}{structure of multi-level clusters.
`nested` (default) or `cross` for cross-classified}

\item{formula}{a string containing a string formula with terms and parameters}

\item{beta}{fixed effect coefficient vector. Overrides parameters in formula.}

\item{varcorr}{random effect variances-covariances.
A list of matrices, one for each cluster.
Overrides parameters in formula.}
}
\value{
an object of class \code{dataframe}
}
\description{
Create a dataset based on a data generating model passed
with a parametrized formula or a set of parameter matrices
}
\details{
more to come
}
\author{
Marcello Gallucci
}
