% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{make_sample}
\alias{make_sample}
\title{Create a model with required parameters}
\usage{
make_sample(
  vars,
  clusters,
  structure,
  formula = NULL,
  fixed = NULL,
  random = NULL,
  factors = "variable",
  empirical = TRUE
)
}
\arguments{
\item{formula}{model formula. It can be a character vector containing
a string formula with terms and parameters (see details). Alternatively,}

\item{data}{a data.frame containing the variables used in the formula.}

\item{beta}{fixed effects coefficients vector. Overrides parameters in formula.}

\item{varcorr}{random effect variances-covariances.
A list of matrices, one for each cluster.
Overrides parameters in formula.}
}
\description{
Create a model with fixed and random effects defined
by the user.
}
