% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelpower.R
\name{modelpower.lmerMod}
\alias{modelpower.lmerMod}
\title{Compute power mixed models}
\usage{
\method{modelpower}{lmerMod}(
  model,
  alpha = 0.05,
  method = "test",
  nsim = 100,
  test = "t",
  df = "sat",
  effects = c("fixed"),
  seed = NULL,
  parallel = TRUE
)
}
\arguments{
\item{model}{of class \code{\link[lme4]{lmerMod}}}

\item{alpha}{significance level (Type I error probability)}

\item{method}{`test` (default) computes power based on the t-test and
model degrees fo freedom. `sim` computes power based on simulations.}

\item{nsim}{Number of simulations for `method="sim"`.Default is 100 to allow trying
out the function. For accurate results, at least 1000 is recommended.}

\item{effects}{a character string or character vector, describing for which
type of coefficients the power is required.
It could be \code{fixed} (default),\code{random} or both (\code{c("fixed","random")}).}

\item{seed}{optional argument to \code{\link[base]{set.seed}}}

\item{parallel}{TRUE (default) of FALSE. Whether the simulations are run
in parallel or in sequence. If the required number of
simulations (`nsim`) is equal or less than 10, FALSE is
enforced. Parallel computation is implemented based on \code{\link[future]} package.}
}
\value{
an object of class \code{\link[paml]{pamlpowermix}}
}
\description{
Compute the power of a mixed model of the class
\code{\link[lme4]{lmerMod}} or \code{\link[lmerTest]{lmerModLmerTest-class}}
}
\details{
Exactly one of the parameters n,
 power, alpha must be passed as NULL, and that parameter
 is determined from the others.
 Notice that alpha has non-NULL default so NULL must be
 explicitly passed if you want it computed.
}
\author{
Marcello Gallucci
}
