% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelpower.R
\name{modelpower}
\alias{modelpower}
\title{Compute the power of a linear model}
\usage{
modelpower(
  model,
  alpha,
  method = "test",
  nsim = 100,
  parallel = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{model}{A model of of class \code{\link[lme4]{lmerMod}}}

\item{alpha}{Significance level (Type I error probability)}

\item{...}{additional options that depends on the class of model
passed to the function.}
}
\value{
an object of class \code{pamlpower*},
        depending on the class of the model that is evaluated
}
\description{
Compute power of parameters tests linear models
}
\details{
Exactly one of the parameters n,
 power, alpha must be passed as NULL, and that parameter
 is determined from the others.
 Notice that alpha has non-NULL default so NULL must be
 explicitly passed if you want it computed.
}
\author{
Marcello Gallucci
}
