% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requiredn.R
\name{required_n}
\alias{required_n}
\title{Find required N for linear models}
\usage{
required_n(
  model,
  expand,
  direction = "within",
  nsim = 100,
  effects = "fixed",
  target = NULL,
  alpha = 0.05,
  power = 0.8,
  tol = 0.05,
  seed = NULL
)
}
\arguments{
\item{model}{of class \code{\link[lme4]{lmerMod}}}

\item{alpha}{significance level (Type I error probability)}

\item{...}{additional options that depends on the class of model
passed to the function}
}
\value{
an object of class \code{\link[paml]{pamlpowermix}}
}
\description{
Compute required N for several classes of linear models.
}
\details{
Exactly one of the parameters n,
 power, alpha must be passed as NULL, and that parameter
 is determined from the others.
 Notice that alpha has non-NULL default so NULL must be
 explicitly passed if you want it computed.
}
\author{
Marcello Gallucci
}
