% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModelList.R
\name{saveModelList}
\alias{saveModelList}
\title{Create, append, and save Model List}
\usage{
saveModelList(model_filename, algorithm, input_features, predicted_feature,
  variableSelection_method, scale, pca, hyperparameter_list, error_metrics,
  train_validation_ratio, train_test_ratio)
}
\arguments{
\item{model_filename}{String, file name to save the Model List.}

\item{algorithm}{String,
\itemize{
 \item "lm" OLS Linear Regression.
 \item "randomForest" Random Forest.
 \item "gbm" Gradient Boosting Machine.
}}

\item{input_features}{String, vector of input features used in model training.}

\item{predicted_feature}{String, name of the feature being predicted.}

\item{variableSelection_method}{String,
\itemize{
 \item "none" No variable selection, all input features are used.
 \item "lasso' LASSO variable selection.
 \item "randomForest" Random Forest variable selection based on variable importance.
}}

\item{scale}{Logical, indicate whether data was scaled for model training.}

\item{pca}{Logical, indicate whether data was transformed via PCA.}

\item{hyperparameter_list}{List, hyperparameters fed into the model.}

\item{error_metrics}{DataFrame, error metrics that are generated from \code{evaluate_RegressionModel}.}

\item{train_validation_ratio}{Ratio of Train Data to Validation Data.}

\item{train_test_ratio}{Ratio of Train Data and Validation Data to Test Data.}
}
\value{
DataFrame of ModelList after updating.
}
\description{
Create and save the Model List for export & record.
}
\seealso{
\code{\link{evaluate_RegressionModel}}
}
